refactor: extract matchObject
